refactor: api, response request 스크리닝 분리
