+++ src/models/screening/response/detailResponseDto.ts
@@ -1,0 +1,15 @@
+import {TEngCategory} from '@/models/enums/category';
+import {TNonNullScreeningBodyRequest} from '../request/screeningRequestDto';
+
+// 스크리닝 디테일 응답 타입
+export interface IScreeningDetail extends TNonNullScreeningBodyRequest {
+  screeningId: number;
+  private: boolean;
+  bookmarked: boolean;
+  reviewed: boolean;
+}
+
+export interface IScreeningDetailContent
+  extends Omit<IScreeningDetail, 'category'> {
+  category: TEngCategory;
+}
