+++ src/screens/screening/detail/components/detailTitle/index.tsx
@@ -3,9 +3,10 @@ import {View} from 'react-native';
 import Typography from '@/components/typography';
 import palette from '@/styles/theme/color';
 import {TEngCategory} from '@/models/enums/category';
+import {getCategory} from '@/utils/getCategory';
 
 import {detailTitleStyles} from './DetailTitle.style';
-import {getCategory} from '@/utils/getCategory';
+
 interface IDetailTitleProps {
   title: string;
   category: TEngCategory;
