+++ src/screens/screening/home/components/reviewScreeningCarousel/index.tsx
@@ -1,9 +1,10 @@
 import {View} from 'react-native';
 
 import Carousel from '@/components/carousel';
-import {IWeekScreeningData} from '@/models/screening/response';
+import {IWeekScreeningData} from '@/models/screening/response/screeningResponseDto';
 
 import {reviewStyles} from './ReviewScreening.style';
+
 interface IReviewScreeningCarouselProp {
   item: IWeekScreeningData[];
 }
